+++ VERSION.md
@@ -1,1 +1,1 @@
-2.0.0
+1.0.0
