+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = '2019, KomodoPlatform'
 author = 'Roman Sztergbaum Tolg Ay'
 
 # The full version, including alpha/beta/rc tags
-release = '2.0.0'
+release = '1.0.0'
 
 
 # -- General configuration ---------------------------------------------------
