+++ modules/CMakeLists.txt
@@ -3,6 +3,7 @@ add_subdirectory(core)
 add_subdirectory(input)
 add_subdirectory(event)
 add_subdirectory(timer)
+add_subdirectory(transform)
 add_subdirectory(geometry)
 add_subdirectory(graphics)
 add_subdirectory(ecs)
