+++ modules/transform/antara/gaming/transform/antara.transform.tests.cpp
@@ -14,17 +14,5 @@
  *                                                                            *
  ******************************************************************************/
 
-#pragma once
-
-#include <meta/sequence/list.hpp>
-#include <meta/sequence/concat.hpp>
-#include <meta/sequence/flatten.hpp>
-#include "antara/gaming/transform/all.transform.components.hpp"
-#include "antara/gaming/graphics/all.graphics.components.hpp"
-#include "antara/gaming/geometry/all.geometry.components.hpp"
-
-namespace antara::gaming::ecs::component
-{
-    using list_component_list = doom::meta::list<transform::components_list, geometry::components_list, graphics::components_list>;
-    using components_list = doom::meta::flatten<list_component_list>;
-}
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include <doctest/doctest.h>
